apps/events/urls.py
from apps.events.views import EventListView
from django.urls import path
from rest_framework import routers

app_name = 'Events API'

router = routers.SimpleRouter()

urlpatterns = [
    path('', EventListView.as_view(), name='list_events'),
]

urlpatterns += router.urls
