pyproject.toml
[tool.poetry]
name = "spectrum"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Django = {extras = ["argon2"], version = "^4.0"}
celery = "^5.1.2"
django-cors-headers = "^3.8.0"
django-storages = "^1.11.1"
psycopg2-binary = "^2.9.1"
mutagen = "^1.45.1"
redis = "^3.5.3"
djangorestframework-simplejwt = "^4.8.0"
gunicorn = "^20.1.0"
argon2 = "^0.1.10"
django-multiselectfield = "^0.1.12"
drf-spectacular = "^0.24.2"
django-extensions = "^3.2.1"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"
pytest-sugar = "^0.9.5"
pillow = "^9.2.0"
django-debug-toolbar = "^3.7.0"
django-admin-cursor-paginator = "^0.1.3"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
django-simple-history = "^3.1.1"
werkzeug = "^2.2.2"
django-eav2 = "^1.2.3"
requests = "^2.28.1"
transliterate = "^1.10.2"

[tool.poetry.dev-dependencies]
vcrpy = "^4.1.1"
pycodestyle = "^2.7.0"
pylint = "^2.10.2"
black = {version = "^21.8b0", allow-prereleases = true}
tblib = "^1.7.0"
django-extensions = "^3.1.3"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 1  # vertical https://pycqa.github.io/isort/docs/configuration/multi_line_output_modes.html
include_trailing_comma = false
force_grid_wrap = 0
use_parentheses = true
combine_as_imports = true
known_django="django"
sections=['FUTURE', 'STDLIB', 'DJANGO', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
known_first_party = []
known_third_party = [
    "requests",
    "wagtail",
    "mimesis",
    "transliterate",
    "rest_framework",
    "drf_yasg",
    "drf_spectacular",
    # "django_filters",
    "social_core",
    "social_django",
    "rest_framework_simplejwt"
]
